gh-451 Fix for segment markers on text output and data paging.
Data paging was allowing the user to jump beyond the end of a stream
whereby just the XML root elements were displayed. This is now fixed by
adding a constraint to the page offset so that the user cannot jump
beyond the last record. Because data paging assumes that segmented
streams have a header and footer, text streams now include segments
after a header and before a footer, even if neither are added, so that
paging always works correctly regardless of the presence of a header or
footer.