gh-460 Fix stream attributes now always recorded for error streams.
In some instances error streams did not always have stream attributes
added to them for fatal errors. This was because processing was failing
early on during pipeline creation. An error was recorded but stream
attributes were not added to the meta data for the error stream.
Processing now ensures that this happens in all cases.