gh-506 Fix regex escaping in dashboard table filters
In v5 dashboard table filters were enhanced to allow parameters to be
used in include/exclude filters. The implementation included the use of
`\` to escape `$` characters that were not to be considered part of a
parameter reference. This change resulted in regular expressions
requiring `\` being escaped with additional `\` characters. This
escaping has now been removed and instead only `$` chars before `{`
chars need escaping when necessary with double `$$` chars, e.g. use
`$${something` if you actually want `${something` not to be replaced
with a parameter.