Fix package and organisation
Package should not be 'rest'. 'Rest' is an architectural approach for
building services. It doesn't really describe the contents of the
package. I would favour 'api' or maybe 'resources', but that's a bit
long and less commonly understood. It also doesn't include the notion of
the interfaces for the actual resource classes.