Add update API method for data volumes
Needed to @JsonIgnore some fields in FsVolume and FsVolumeState.
These fields are added to the generated JSON by methods, e.g.
getPercentUsed(). But this means the JSON that's generated will fail
to be deserialised back to a POJO. One options is to @JsonIgnore, but then
these potentially useful values don't get sent to the UI. Another option
is to add a field to the class, so they do get deserialised. But that's
a useless field. Yet another option is to use an entirely different POJO
when accepting an update from the client. But this means more POJOs
and more complexity. I've favoured @JsonIgnore because I'm not currently
using these values in the UI.