#1112 You can now copy feeds along with other items
You can now copy feeds along with other items and copies are named
appropriately.

When copying a selection of several items, the dependencies between the
items are altered in the resulting copies so that the copied items work
together as a new set of content.

As part of fixing dependencies when copying items, the dependencies
screen now works correctly and now also shows processor filters.