#1557 Many changes related to process/reprocess data
Process, reprocess, delete and download data functions now provide an
impact summary before a user can proceed with the action.

The process data function in the data browser now provides the option to
process or reprocess data. When selected a user can also choose: the
priority of the process filters that will be created; to set the
priority automatically based on previous filters; set the enabled state.

Reprocessing data no longer has a limitation on how many items can be
reprocessed as it is now implemented by reprocess specific filters.