gh-2526 Fix incorrect input pane in stepper for cooked xml
Changes ReaderRecorder to treat the range end as inclusive or exclusive
depending on the type of parser that is parsing the input data.
This is because of the differences in the way the xml parser and
the dspaser record locations.

Also remove line breaks from the start of the range in the input pane