gh-2787 Add tracker tbl, fix prop updating in UI
The UI refresh was not calling the update of the props from the db
so some nodes were not aware db values had changes. Now it will
update the config from the db but only if the tracker tbl
indicates a change to the config.

Also fixed the refreshing of the UI list so it clears out the local
state before each refresh.

Also changed the cluster effective values to show entries for dead
nodes.