gh-3797 Fix various doc perm issues (WIP)
Fix bug preventing an owner from granting permissions on a document/folder.
Change the document permissions screen to show only users that share at least one group with the user changing the permissions.
This is to prevent lowly users from seeing all users on the system.
Fix problem of ownership being removed from the owner.
Fix other issues with the permission changes not reflecting the changes made in the UI.